requirements.txt
google-cloud-aiplatform==1.95.1
google-cloud-storage==2.19.0
google-genai==1.21.1
gitpython==3.1.40
google-adk==1.14.1
yfinance==0.2.32
psutil==7.0.0
litellm==1.77.1
google-generativeai==0.8.5
python-dotenv==1.1.0

# Test dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
pytest-xdist==3.5.0  # for parallel test execution